# fraudcase_ai/server/app.py
"""FastAPI bridge between the web UI and the FraudCase AI audit agent.

CONTRACT (the Frontend slice codes against exactly these routes):
    POST /api/audit-case        {text}                  -> {case_id}
    GET  /api/events/{case_id}                         -> text/event-stream of AgentEvent
    POST /api/approve/{case_id}  ApprovalDecision      -> {ok: true}
    GET  /api/report/{case_id}                         -> AuditReport
    GET  /healthz                                     -> {status:"ok"}

The Backend/API slice implements the bodies (running the audit agent, relaying the
approval/resume signal, fanning AgentEvents to the SSE stream). Keep the routes/shapes
stable — the Frontend depends on them.
"""

from __future__ import annotations

import asyncio
from typing import AsyncIterator

from fastapi import FastAPI, HTTPException
from fastapi.responses import StreamingResponse

from fraudcase_ai.models import (
    AgentEvent,
    ApprovalDecision,
    AuditReport,
    AskRequest,
    AskResponse,
    EventType,
    AuditCaseRequest,
    AuditCaseStarted,
)
from fraudcase_ai.server import insights
from fraudcase_ai.server.events import to_sse
from fraudcase_ai.server.runner import FakeRunner
from fraudcase_ai.config import get_settings
from fraudcase_ai.server.store import AuditCaseStore, audit_case_store as _default_store

app = FastAPI(title="FraudCase AI")

# Allow tests to inject a custom store by replacing app.state.store.
# The default is the module-level singleton.
app.state.store = _default_store


def _get_store() -> AuditCaseStore:
    return app.state.store  # type: ignore[attr-defined]


# --------------------------------------------------------------------------- #
# Background task: drive the runner and feed the event queue
# --------------------------------------------------------------------------- #

async def _process_case(case_id: str, store: AuditCaseStore) -> None:
    """Background task: iterate AgentEvents from the runner into the queue."""
    record = store.get_case(case_id)
    if record is None:
        return

    try:
        async for event in record.runner.run(case_id, record.audit_case):
            store.apply_event(event)
            await record.queue.put(event)
            # Cache the AuditReport when REPORT_READY fires so /api/report works.
            if event.type == EventType.REPORT_READY:
                report_data = event.data.get("report")
                if report_data:
                    report = AuditReport.model_validate(report_data)
                    store.set_report(case_id, report)
    except Exception as exc:  # noqa: BLE001
        err_event = AgentEvent(
            case_id=case_id,
            type=EventType.ERROR,
            data={"error": str(exc)},
        )
        await record.queue.put(err_event)
        done_event = AgentEvent(case_id=case_id, type=EventType.DONE, data={})
        await record.queue.put(done_event)
    finally:
        # Sentinel: None signals the SSE generator to close the stream.
        await record.queue.put(None)


# --------------------------------------------------------------------------- #
# SSE generator
# --------------------------------------------------------------------------- #

async def _sse_generator(case_id: str, store: AuditCaseStore) -> AsyncIterator[str]:
    record = store.get_case(case_id)
    if record is None:
        return

    while True:
        event: AgentEvent | None = await record.queue.get()
        if event is None:
            break
        yield to_sse(event)
        if event.type == EventType.DONE:
            break


# --------------------------------------------------------------------------- #
# Routes
# --------------------------------------------------------------------------- #

@app.get("/healthz")
def healthz() -> dict:
    return {"status": "ok"}


@app.get("/api/status")
def status() -> dict:
    return insights.get_status()


@app.get("/api/stats")
def stats() -> dict:
    return insights.get_stats()


@app.post("/api/ask", response_model=AskResponse)
def ask(body: AskRequest) -> AskResponse:
    store = _get_store()
    report = store.get_report(body.case_id) if body.case_id else store.newest_report()
    return insights.answer_question(body.question, report, body.invoice_context)


@app.post("/api/audit-case", response_model=AuditCaseStarted)
async def start_audit_case(body: AuditCaseRequest) -> AuditCaseStarted:
    """Start a new audit case. Returns a case_id immediately.

    The runner is launched as a background task and feeds AgentEvents into the
    audit case queue, which the /api/events/{case_id} SSE stream consumes.
    """
    store = _get_store()
    settings = get_settings()
    if settings.use_mocks:
        runner = FakeRunner()
    else:
        from fraudcase_ai.server.real_runner import RealRunner

        runner = RealRunner()  # live: Gemini + Atlas vector search
    record = store.create_case(body, runner)
    asyncio.create_task(_process_case(record.case_id, store))
    return AuditCaseStarted(case_id=record.case_id)


@app.post("/api/maestro/audit-cases", response_model=AuditCaseStarted)
async def maestro_start_audit_case(body: AuditCaseRequest) -> AuditCaseStarted:
    """UiPath Maestro service-task entrypoint for opening an audit case."""
    return await start_audit_case(body)


@app.get("/api/maestro/audit-cases/{case_id}")
def maestro_get_audit_case(case_id: str) -> dict:
    """Return Maestro-facing state for an audit case."""
    store = _get_store()
    try:
        return store.case_summary(case_id)
    except KeyError:
        raise HTTPException(status_code=404, detail=f"Audit case {case_id!r} not found")


@app.post("/api/maestro/audit-cases/{case_id}/decisions")
async def maestro_decide_audit_case(case_id: str, body: ApprovalDecision) -> dict:
    """UiPath Maestro human-task callback for Gate 1 and Gate 2 decisions."""
    return await approve(case_id, body)


@app.get("/api/events/{case_id}")
async def stream_events(case_id: str) -> StreamingResponse:
    """Stream AgentEvents as Server-Sent-Events until DONE."""
    store = _get_store()
    if store.get_case(case_id) is None:
        raise HTTPException(status_code=404, detail=f"Audit case {case_id!r} not found")
    return StreamingResponse(
        _sse_generator(case_id, store),
        media_type="text/event-stream",
        headers={
            "Cache-Control": "no-cache",
            "X-Accel-Buffering": "no",
        },
    )


@app.post("/api/approve/{case_id}")
async def approve(case_id: str, body: ApprovalDecision) -> dict:
    """Deliver an approval decision to resume a paused audit case."""
    store = _get_store()
    try:
        await store.push_approval(case_id, body)
    except KeyError:
        raise HTTPException(status_code=404, detail=f"Audit case {case_id!r} not found")
    return {"ok": True}


@app.get("/api/report/{case_id}", response_model=AuditReport)
def get_report(case_id: str) -> AuditReport:
    """Return the final AuditReport for a completed audit case."""
    store = _get_store()
    report = store.get_report(case_id)
    if report is None:
        if store.get_case(case_id) is None:
            raise HTTPException(status_code=404, detail=f"Audit case {case_id!r} not found")
        raise HTTPException(status_code=404, detail="Report not yet available")
    return report


# --------------------------------------------------------------------------- #
# Serve the web UI (mounted last so it doesn't shadow /api routes)
# --------------------------------------------------------------------------- #
from pathlib import Path  # noqa: E402

from fastapi.staticfiles import StaticFiles  # noqa: E402

_WEB_DIR = Path(__file__).resolve().parent.parent / "web"
if _WEB_DIR.is_dir():
    app.mount("/", StaticFiles(directory=str(_WEB_DIR), html=True), name="web")